requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
gunicorn==21.2.0
asgiref==3.8.1

langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.60
langchain-openai==0.3.17
qdrant-client>=1.7.0
langchain-qdrant>=0.1.0
langgraph==0.2.48
langgraph-checkpoint==2.0.4
langgraph-sdk==0.1.36
langsmith==0.1.145

redis==5.0.1
hiredis==2.3.2
flask-limiter==3.5.0

python-dotenv==1.0.1
websockets==13.1
pydantic==2.10.5
pydantic_core==2.27.2
pydantic-settings==2.6.1
httpx==0.27.0

wsproto==1.2.0
yarl==1.17.1
pydub==0.25.1
streamlit==1.40.0
pdfplumber>=0.10.0